feat: move prev components into feature for viewer
