+++ src/ui/features/viewer/components/scene-manager/view-card/CreateViewCard.css
