+++ src/ui/features/viewer/components/scene-manager/view-card/ViewCard.css
